moved chess-related code into one file
