+++ docs/decorator/script.js
@@ -2833,6 +2833,52 @@
     },
     SvgElement: function SvgElement() {
     },
+    ChessPiece$(colour, $name, moveStrategy) {
+      var t1 = new A.ChessPiece(moveStrategy, colour, $name);
+      t1.__GamePiece_src_A = "./assets/chess/" + $name + "_" + colour + ".png";
+      return t1;
+    },
+    EmptyPiece$(iPosition, jPosition) {
+      var t1 = new A.EmptyPiece(new A.NoMovement(), "empty", "empty");
+      t1.__GamePiece_src_A = "./assets/chess/empty_empty.png";
+      t1.i = iPosition;
+      t1.j = jPosition;
+      return t1;
+    },
+    ChessPiece: function ChessPiece(t0, t1, t2) {
+      var _ = this;
+      _.j = _.i = 0;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.__GamePiece_src_A = $;
+    },
+    EmptyPiece: function EmptyPiece(t0, t1, t2) {
+      var _ = this;
+      _.j = _.i = 0;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.__GamePiece_src_A = $;
+    },
+    ChessGame: function ChessGame(t0, t1) {
+      var _ = this;
+      _.turnCount = 0;
+      _.__ChessGame_board_AI = $;
+      _.activePiece = t0;
+      _.__ChessGame_view_A = $;
+      _.container = t1;
+    },
+    ChessBoardView: function ChessBoardView(t0, t1) {
+      this.container = t0;
+      this.game = t1;
+    },
+    ChessBoardView_createTile_closure: function ChessBoardView_createTile_closure(t0, t1) {
+      this.$this = t0;
+      this.piece = t1;
+    },
     BoardWithPieces: function BoardWithPieces() {
     },
     BoardWithPawns: function BoardWithPawns(t0) {
@@ -2853,56 +2899,24 @@
     BoardWithQueens: function BoardWithQueens(t0) {
       this.base = t0;
     },
-    ChessBoardView: function ChessBoardView(t0, t1) {
-      this.container = t0;
-      this.game = t1;
-    },
-    ChessBoardView_createTile_closure: function ChessBoardView_createTile_closure(t0, t1) {
-      this.$this = t0;
-      this.piece = t1;
-    },
-    ChessGame: function ChessGame(t0, t1) {
-      var _ = this;
-      _.turnCount = 0;
-      _.__ChessGame_board_AI = $;
-      _.activePiece = t0;
-      _.__ChessGame_view_A = $;
-      _.container = t1;
-    },
     ChequeredBoard: function ChequeredBoard(t0) {
       this.pieces = t0;
     },
-    ChessPiece$(colour, $name, moveStrategy) {
-      var t1 = new A.ChessPiece(moveStrategy, colour, $name);
-      t1.__GamePiece_src_A = "./assets/chess/" + $name + "_" + colour + ".png";
-      return t1;
+    NoMovement: function NoMovement() {
     },
-    EmptyPiece$(iPosition, jPosition) {
-      var t1 = new A.EmptyPiece(new A.NoMovement(), "empty", "empty");
-      t1.__GamePiece_src_A = "./assets/chess/empty_empty.png";
-      t1.i = iPosition;
-      t1.j = jPosition;
-      return t1;
+    PawnMovement: function PawnMovement() {
+    },
+    KnightMovement: function KnightMovement() {
+    },
+    BishopMovement: function BishopMovement() {
+    },
+    RookMovement: function RookMovement() {
+    },
+    QueenMovement: function QueenMovement() {
+    },
+    KingMovement: function KingMovement() {
     },
     GamePiece: function GamePiece() {
-    },
-    ChessPiece: function ChessPiece(t0, t1, t2) {
-      var _ = this;
-      _.j = _.i = 0;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.__GamePiece_src_A = $;
-    },
-    EmptyPiece: function EmptyPiece(t0, t1, t2) {
-      var _ = this;
-      _.j = _.i = 0;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.__GamePiece_src_A = $;
     },
     main() {
       var game, initState,
@@ -2946,20 +2960,6 @@
     },
     DecoratorCheckbox_armCheckbox_closure: function DecoratorCheckbox_armCheckbox_closure(t0) {
       this.$this = t0;
-    },
-    NoMovement: function NoMovement() {
-    },
-    PawnMovement: function PawnMovement() {
-    },
-    KnightMovement: function KnightMovement() {
-    },
-    BishopMovement: function BishopMovement() {
-    },
-    RookMovement: function RookMovement() {
-    },
-    QueenMovement: function QueenMovement() {
-    },
-    KingMovement: function KingMovement() {
     },
     throwLateFieldNI(fieldName) {
       return A.throwExpression(A.LateError$fieldNI(fieldName));
@@ -4401,6 +4401,175 @@
       return new A.FilteredElementList(new A._ChildNodeListLazy(receiver));
     }
   };
+  A.ChessPiece.prototype = {
+    canCapture$3(board, i, j) {
+      var target;
+      if (this.validCoords$2(i, j)) {
+        target = board.getPiece$2(i, j);
+        if (!(target instanceof A.EmptyPiece) && target.colour !== this.colour)
+          return target.threatened = true;
+      }
+      return false;
+    },
+    canMove$3(board, i, j) {
+      var target;
+      if (this.validCoords$2(i, j)) {
+        target = board.getPiece$2(i, j);
+        if (target instanceof A.EmptyPiece)
+          return target.threatened = true;
+      }
+      return false;
+    },
+    validCoords$2(i, j) {
+      var t1 = 0 <= i && i < 8;
+      return B.JSBool_methods.$and(t1, 0 <= j && j < 8);
+    }
+  };
+  A.EmptyPiece.prototype = {};
+  A.ChessGame.prototype = {
+    get$board() {
+      var t1,
+        value = this.__ChessGame_board_AI;
+      if (value === $) {
+        t1 = J.JSArray_JSArray$growable(0, type$.List_ChessPiece);
+        value = this.__ChessGame_board_AI = new A.ChequeredBoard(t1);
+      }
+      return value;
+    },
+    submitMove$2(i, j) {
+      var t1, piece, _this = this;
+      if (_this.get$board().getPiece$2(i, j).threatened) {
+        t1 = _this.activePiece;
+        _this.get$board().removePiece$2(t1.i, t1.j);
+        _this.get$board().removePiece$2(i, j);
+        _this.get$board().placePiece$3(t1, i, j);
+        t1.hasMoved = true;
+        ++_this.turnCount;
+        _this.activePiece = A.EmptyPiece$(0, 0);
+        _this.clearMoveOptions$0();
+        _this.refreshView$0();
+        return;
+      }
+      _this.clearMoveOptions$0();
+      _this.activePiece = A.EmptyPiece$(0, 0);
+      piece = _this.get$board().getPiece$2(i, j);
+      t1 = B.JSInt_methods.$mod(_this.turnCount, 2) === 0 ? "w" : "b";
+      if (piece.colour === t1) {
+        piece.moveStrategy.move$2(_this.get$board(), piece);
+        _this.activePiece = piece;
+      }
+      _this.refreshView$0();
+    },
+    refreshView$0() {
+      var t1 = this.__ChessGame_view_A;
+      t1 === $ && A.throwLateFieldNI("view");
+      t1.displayBoard$1(this.get$board().getBoardState$0());
+    },
+    clearMoveOptions$0() {
+      var t1, i, j, value, t2;
+      for (t1 = type$.JSArray_List_ChessPiece, i = 0; i < 8; ++i)
+        for (j = 0; j < 8; ++j) {
+          value = this.__ChessGame_board_AI;
+          if (value === $) {
+            t2 = A._setArrayType(new Array(0), t1);
+            value = this.__ChessGame_board_AI = new A.ChequeredBoard(t2);
+          }
+          value.getPiece$2(i, j).threatened = false;
+        }
+    },
+    setupPieces$1(state) {
+      var t1, _this = this;
+      type$.List_bool._as(state);
+      t1 = J.JSArray_JSArray$growable(0, type$.List_ChessPiece);
+      _this.__ChessGame_board_AI = new A.ChequeredBoard(t1);
+      if (0 >= state.length)
+        return A.ioore(state, 0);
+      if (state[0])
+        _this.__ChessGame_board_AI = new A.BoardWithPawns(_this.get$board());
+      if (1 >= state.length)
+        return A.ioore(state, 1);
+      if (state[1])
+        _this.__ChessGame_board_AI = new A.BoardWithBishops(_this.get$board());
+      if (2 >= state.length)
+        return A.ioore(state, 2);
+      if (state[2])
+        _this.__ChessGame_board_AI = new A.BoardWithKnights(_this.get$board());
+      if (3 >= state.length)
+        return A.ioore(state, 3);
+      if (state[3])
+        _this.__ChessGame_board_AI = new A.BoardWithRooks(_this.get$board());
+      if (4 >= state.length)
+        return A.ioore(state, 4);
+      if (state[4])
+        _this.__ChessGame_board_AI = new A.BoardWithQueens(_this.get$board());
+      if (5 >= state.length)
+        return A.ioore(state, 5);
+      if (state[5])
+        _this.__ChessGame_board_AI = new A.BoardWithKings(_this.get$board());
+      _this.get$board().setupPieces$0();
+      _this.refreshView$0();
+    }
+  };
+  A.ChessBoardView.prototype = {
+    displayBoard$1(boardstate) {
+      var t1, t2, t3, t4, _i, rowOfPieces, t5, row, t6, t7, t8, tile, t9, img, t10, element, subtype;
+      type$.List_List_ChessPiece._as(boardstate);
+      t1 = this.container;
+      t2 = J.getInterceptor$x(t1);
+      t2.get$children(t1).clear$0(0);
+      for (t3 = boardstate.length, t4 = type$.ImageElement, _i = 0; _i < boardstate.length; boardstate.length === t3 || (0, A.throwConcurrentModificationError)(boardstate), ++_i) {
+        rowOfPieces = boardstate[_i];
+        t5 = document;
+        row = t5.createElement("div");
+        t6 = J.getInterceptor$x(row);
+        t6.get$classes(row).add$1(0, "board-row");
+        for (t7 = B.JSArray_methods.get$iterator(rowOfPieces); t7.moveNext$0();) {
+          t8 = t7.get$current();
+          tile = this.createTile$1(t8);
+          t9 = t8 instanceof A.EmptyPiece;
+          if (!t9) {
+            img = t5.createElement("img");
+            J.get$classes$x(img).add$1(0, "piece-img");
+            if (t4._is(img)) {
+              t10 = t8.__GamePiece_src_A;
+              t10 === $ && A.throwLateFieldNI("src");
+              B.ImageElement_methods.set$src(img, t10);
+            }
+            J.get$children$x(tile).add$1(0, img);
+          }
+          if (t8.threatened) {
+            element = t5.createElement("div");
+            t8 = J.getInterceptor$x(element);
+            t8.get$classes(element).add$1(0, "marker");
+            subtype = t9 ? "dot" : "circle";
+            t8.get$classes(element).add$1(0, subtype);
+            J.get$children$x(tile).add$1(0, element);
+          }
+          t6.get$children(row).add$1(0, tile);
+        }
+        t2.get$children(t1).add$1(0, row);
+      }
+    },
+    createTile$1(piece) {
+      var tile = document.createElement("div"),
+        t1 = J.getInterceptor$x(tile);
+      t1.get$classes(tile).add$1(0, "chess-tile");
+      if (B.JSInt_methods.$mod(piece.i + piece.j, 2) !== 0)
+        t1.get$classes(tile).add$1(0, "dark");
+      t1.addEventListener$2(tile, "click", new A.ChessBoardView_createTile_closure(this, piece));
+      return tile;
+    },
+    $isChessView: 1
+  };
+  A.ChessBoardView_createTile_closure.prototype = {
+    call$1($event) {
+      var t1;
+      type$.Event._as($event);
+      t1 = this.piece;
+      this.$this.game.submitMove$2(t1.i, t1.j);
+    },
+    $signature: 0
+  };
   A.BoardWithPieces.prototype = {
     removePiece$2(i, j) {
       this.base.removePiece$2(i, j);
@@ -4490,150 +4659,6 @@
       this.base.placePiece$3(A.ChessPiece$(colour, "queen", new A.QueenMovement()), i, j);
     }
   };
-  A.ChessBoardView.prototype = {
-    displayBoard$1(boardstate) {
-      var t1, t2, t3, t4, _i, rowOfPieces, t5, row, t6, t7, t8, tile, t9, img, t10, element, subtype;
-      type$.List_List_ChessPiece._as(boardstate);
-      t1 = this.container;
-      t2 = J.getInterceptor$x(t1);
-      t2.get$children(t1).clear$0(0);
-      for (t3 = boardstate.length, t4 = type$.ImageElement, _i = 0; _i < boardstate.length; boardstate.length === t3 || (0, A.throwConcurrentModificationError)(boardstate), ++_i) {
-        rowOfPieces = boardstate[_i];
-        t5 = document;
-        row = t5.createElement("div");
-        t6 = J.getInterceptor$x(row);
-        t6.get$classes(row).add$1(0, "board-row");
-        for (t7 = B.JSArray_methods.get$iterator(rowOfPieces); t7.moveNext$0();) {
-          t8 = t7.get$current();
-          tile = this.createTile$1(t8);
-          t9 = t8 instanceof A.EmptyPiece;
-          if (!t9) {
-            img = t5.createElement("img");
-            J.get$classes$x(img).add$1(0, "piece-img");
-            if (t4._is(img)) {
-              t10 = t8.__GamePiece_src_A;
-              t10 === $ && A.throwLateFieldNI("src");
-              B.ImageElement_methods.set$src(img, t10);
-            }
-            J.get$children$x(tile).add$1(0, img);
-          }
-          if (t8.threatened) {
-            element = t5.createElement("div");
-            t8 = J.getInterceptor$x(element);
-            t8.get$classes(element).add$1(0, "marker");
-            subtype = t9 ? "dot" : "circle";
-            t8.get$classes(element).add$1(0, subtype);
-            J.get$children$x(tile).add$1(0, element);
-          }
-          t6.get$children(row).add$1(0, tile);
-        }
-        t2.get$children(t1).add$1(0, row);
-      }
-    },
-    createTile$1(piece) {
-      var tile = document.createElement("div"),
-        t1 = J.getInterceptor$x(tile);
-      t1.get$classes(tile).add$1(0, "chess-tile");
-      if (B.JSInt_methods.$mod(piece.i + piece.j, 2) !== 0)
-        t1.get$classes(tile).add$1(0, "dark");
-      t1.addEventListener$2(tile, "click", new A.ChessBoardView_createTile_closure(this, piece));
-      return tile;
-    },
-    $isChessView: 1
-  };
-  A.ChessBoardView_createTile_closure.prototype = {
-    call$1($event) {
-      var t1;
-      type$.Event._as($event);
-      t1 = this.piece;
-      this.$this.game.submitMove$2(t1.i, t1.j);
-    },
-    $signature: 0
-  };
-  A.ChessGame.prototype = {
-    get$board() {
-      var t1,
-        value = this.__ChessGame_board_AI;
-      if (value === $) {
-        t1 = J.JSArray_JSArray$growable(0, type$.List_ChessPiece);
-        value = this.__ChessGame_board_AI = new A.ChequeredBoard(t1);
-      }
-      return value;
-    },
-    submitMove$2(i, j) {
-      var t1, piece, _this = this;
-      if (_this.get$board().getPiece$2(i, j).threatened) {
-        t1 = _this.activePiece;
-        _this.get$board().removePiece$2(t1.i, t1.j);
-        _this.get$board().removePiece$2(i, j);
-        _this.get$board().placePiece$3(t1, i, j);
-        t1.hasMoved = true;
-        ++_this.turnCount;
-        _this.activePiece = A.EmptyPiece$(0, 0);
-        _this.clearMoveOptions$0();
-        _this.refreshView$0();
-        return;
-      }
-      _this.clearMoveOptions$0();
-      _this.activePiece = A.EmptyPiece$(0, 0);
-      piece = _this.get$board().getPiece$2(i, j);
-      t1 = B.JSInt_methods.$mod(_this.turnCount, 2) === 0 ? "w" : "b";
-      if (piece.colour === t1) {
-        piece.moveStrategy.move$2(_this.get$board(), piece);
-        _this.activePiece = piece;
-      }
-      _this.refreshView$0();
-    },
-    refreshView$0() {
-      var t1 = this.__ChessGame_view_A;
-      t1 === $ && A.throwLateFieldNI("view");
-      t1.displayBoard$1(this.get$board().getBoardState$0());
-    },
-    clearMoveOptions$0() {
-      var t1, i, j, value, t2;
-      for (t1 = type$.JSArray_List_ChessPiece, i = 0; i < 8; ++i)
-        for (j = 0; j < 8; ++j) {
-          value = this.__ChessGame_board_AI;
-          if (value === $) {
-            t2 = A._setArrayType(new Array(0), t1);
-            value = this.__ChessGame_board_AI = new A.ChequeredBoard(t2);
-          }
-          value.getPiece$2(i, j).threatened = false;
-        }
-    },
-    setupPieces$1(state) {
-      var t1, _this = this;
-      type$.List_bool._as(state);
-      t1 = J.JSArray_JSArray$growable(0, type$.List_ChessPiece);
-      _this.__ChessGame_board_AI = new A.ChequeredBoard(t1);
-      if (0 >= state.length)
-        return A.ioore(state, 0);
-      if (state[0])
-        _this.__ChessGame_board_AI = new A.BoardWithPawns(_this.get$board());
-      if (1 >= state.length)
-        return A.ioore(state, 1);
-      if (state[1])
-        _this.__ChessGame_board_AI = new A.BoardWithBishops(_this.get$board());
-      if (2 >= state.length)
-        return A.ioore(state, 2);
-      if (state[2])
-        _this.__ChessGame_board_AI = new A.BoardWithKnights(_this.get$board());
-      if (3 >= state.length)
-        return A.ioore(state, 3);
-      if (state[3])
-        _this.__ChessGame_board_AI = new A.BoardWithRooks(_this.get$board());
-      if (4 >= state.length)
-        return A.ioore(state, 4);
-      if (state[4])
-        _this.__ChessGame_board_AI = new A.BoardWithQueens(_this.get$board());
-      if (5 >= state.length)
-        return A.ioore(state, 5);
-      if (state[5])
-        _this.__ChessGame_board_AI = new A.BoardWithKings(_this.get$board());
-      _this.get$board().setupPieces$0();
-      _this.refreshView$0();
-    }
-  };
   A.ChequeredBoard.prototype = {
     setupPieces$0() {
       var t1, t2, i, row, j;
@@ -4671,86 +4696,6 @@
       return t1[j];
     },
     $isChessBoard: 1
-  };
-  A.GamePiece.prototype = {};
-  A.ChessPiece.prototype = {
-    canCapture$3(board, i, j) {
-      var target;
-      if (this.validCoords$2(i, j)) {
-        target = board.getPiece$2(i, j);
-        if (!(target instanceof A.EmptyPiece) && target.colour !== this.colour)
-          return target.threatened = true;
-      }
-      return false;
-    },
-    canMove$3(board, i, j) {
-      var target;
-      if (this.validCoords$2(i, j)) {
-        target = board.getPiece$2(i, j);
-        if (target instanceof A.EmptyPiece)
-          return target.threatened = true;
-      }
-      return false;
-    },
-    validCoords$2(i, j) {
-      var t1 = 0 <= i && i < 8;
-      return B.JSBool_methods.$and(t1, 0 <= j && j < 8);
-    }
-  };
-  A.EmptyPiece.prototype = {};
-  A.DecoratorDemo.prototype = {
-    addCheckbox$1(piece) {
-      var cbox, t4, t5, label,
-        t1 = document,
-        t2 = t1.createElement("div"),
-        box = new A.DecoratorCheckbox(t2, this, piece),
-        t3 = J.getInterceptor$x(t2);
-      t3.get$classes(t2).add$1(0, "deco-cbox");
-      cbox = t1.createElement("input");
-      t4 = piece + "-cbox";
-      cbox.id = t4;
-      J.get$classes$x(cbox).add$1(0, "cbox");
-      t5 = type$.InputElement._is(cbox);
-      if (t5) {
-        B.InputElement_methods.set$type(cbox, "checkbox");
-        B.InputElement_methods.set$checked(cbox, true);
-      }
-      t3.get$children(t2).add$1(0, cbox);
-      label = t1.createElement("label");
-      J.get$classes$x(label).add$1(0, "cbox-label");
-      if (type$.LabelElement._is(label)) {
-        label.htmlFor = t4;
-        label.innerText = piece + "s";
-      }
-      t3.get$children(t2).add$1(0, label);
-      if (t5)
-        box.armCheckbox$1(cbox);
-      B.JSArray_methods.add$1(this.checkboxes, box);
-      J.get$children$x(this.sideTray).add$1(0, t2);
-    },
-    rebuildGame$0() {
-      var t1, i,
-        state = J.JSArray_JSArray$growable(0, type$.bool);
-      for (t1 = this.checkboxes, i = 0; i < t1.length; ++i)
-        B.JSArray_methods.add$1(state, t1[i].checked);
-      this.game.setupPieces$1(state);
-    },
-    $isDemo: 1
-  };
-  A.DecoratorCheckbox.prototype = {
-    armCheckbox$1(cbox) {
-      B.InputElement_methods.addEventListener$2(cbox, "input", new A.DecoratorCheckbox_armCheckbox_closure(this));
-    }
-  };
-  A.DecoratorCheckbox_armCheckbox_closure.prototype = {
-    call$1($event) {
-      var t1;
-      type$.Event._as($event);
-      t1 = this.$this;
-      t1.checked = !t1.checked;
-      t1.demo.rebuildGame$0();
-    },
-    $signature: 0
   };
   A.NoMovement.prototype = {
     move$2(board, piece) {
@@ -4876,6 +4821,61 @@
     },
     $isMovementStrategy: 1
   };
+  A.GamePiece.prototype = {};
+  A.DecoratorDemo.prototype = {
+    addCheckbox$1(piece) {
+      var cbox, t4, t5, label,
+        t1 = document,
+        t2 = t1.createElement("div"),
+        box = new A.DecoratorCheckbox(t2, this, piece),
+        t3 = J.getInterceptor$x(t2);
+      t3.get$classes(t2).add$1(0, "deco-cbox");
+      cbox = t1.createElement("input");
+      t4 = piece + "-cbox";
+      cbox.id = t4;
+      J.get$classes$x(cbox).add$1(0, "cbox");
+      t5 = type$.InputElement._is(cbox);
+      if (t5) {
+        B.InputElement_methods.set$type(cbox, "checkbox");
+        B.InputElement_methods.set$checked(cbox, true);
+      }
+      t3.get$children(t2).add$1(0, cbox);
+      label = t1.createElement("label");
+      J.get$classes$x(label).add$1(0, "cbox-label");
+      if (type$.LabelElement._is(label)) {
+        label.htmlFor = t4;
+        label.innerText = piece + "s";
+      }
+      t3.get$children(t2).add$1(0, label);
+      if (t5)
+        box.armCheckbox$1(cbox);
+      B.JSArray_methods.add$1(this.checkboxes, box);
+      J.get$children$x(this.sideTray).add$1(0, t2);
+    },
+    rebuildGame$0() {
+      var t1, i,
+        state = J.JSArray_JSArray$growable(0, type$.bool);
+      for (t1 = this.checkboxes, i = 0; i < t1.length; ++i)
+        B.JSArray_methods.add$1(state, t1[i].checked);
+      this.game.setupPieces$1(state);
+    },
+    $isDemo: 1
+  };
+  A.DecoratorCheckbox.prototype = {
+    armCheckbox$1(cbox) {
+      B.InputElement_methods.addEventListener$2(cbox, "input", new A.DecoratorCheckbox_armCheckbox_closure(this));
+    }
+  };
+  A.DecoratorCheckbox_armCheckbox_closure.prototype = {
+    call$1($event) {
+      var t1;
+      type$.Event._as($event);
+      t1 = this.$this;
+      t1.checked = !t1.checked;
+      t1.demo.rebuildGame$0();
+    },
+    $signature: 0
+  };
   (function aliases() {
     var _ = J.Interceptor.prototype;
     _.super$Interceptor$toString = _.toString$0;
@@ -4887,7 +4887,7 @@
       _inherit = hunkHelpers.inherit,
       _inheritMany = hunkHelpers.inheritMany;
     _inherit(A.Object, null);
-    _inheritMany(A.Object, [A.JS_CONST, J.Interceptor, J.ArrayIterator, A.Error, A.ListIterator, A.Iterable, A.MappedIterator, A.WhereIterator, A.Closure, A.JSSyntaxRegExp, A.Rti, A._FunctionParameters, A._Type, A.SetBase, A._LinkedHashSetCell, A._LinkedHashSetIterator, A.ListBase, A._Exception, A.FormatException, A.Null, A.StringBuffer, A.ImmutableListMixin, A.FixedSizeListIterator, A.BoardWithPieces, A.ChessBoardView, A.ChessGame, A.ChequeredBoard, A.GamePiece, A.DecoratorDemo, A.DecoratorCheckbox, A.NoMovement, A.PawnMovement, A.KnightMovement, A.BishopMovement, A.RookMovement, A.QueenMovement, A.KingMovement]);
+    _inheritMany(A.Object, [A.JS_CONST, J.Interceptor, J.ArrayIterator, A.Error, A.ListIterator, A.Iterable, A.MappedIterator, A.WhereIterator, A.Closure, A.JSSyntaxRegExp, A.Rti, A._FunctionParameters, A._Type, A.SetBase, A._LinkedHashSetCell, A._LinkedHashSetIterator, A.ListBase, A._Exception, A.FormatException, A.Null, A.StringBuffer, A.ImmutableListMixin, A.FixedSizeListIterator, A.GamePiece, A.ChessGame, A.ChessBoardView, A.BoardWithPieces, A.ChequeredBoard, A.NoMovement, A.PawnMovement, A.KnightMovement, A.BishopMovement, A.RookMovement, A.QueenMovement, A.KingMovement, A.DecoratorDemo, A.DecoratorCheckbox]);
     _inheritMany(J.Interceptor, [J.JSBool, J.JSNull, J.JavaScriptObject, J.JSNumber, J.JSString]);
     _inheritMany(J.JavaScriptObject, [J.LegacyJavaScriptObject, J.JSArray, A.EventTarget, A.DomException, A.DomTokenList, A.Event, A._HtmlCollection_JavaScriptObject_ListMixin, A._NodeList_JavaScriptObject_ListMixin, A.__NamedNodeMap_JavaScriptObject_ListMixin]);
     _inheritMany(J.LegacyJavaScriptObject, [J.PlainJavaScriptObject, J.UnknownJavaScriptObject, J.JavaScriptFunction]);
@@ -4915,9 +4915,9 @@
     _inherit(A.__NamedNodeMap_JavaScriptObject_ListMixin_ImmutableListMixin, A.__NamedNodeMap_JavaScriptObject_ListMixin);
     _inherit(A._NamedNodeMap, A.__NamedNodeMap_JavaScriptObject_ListMixin_ImmutableListMixin);
     _inheritMany(A.CssClassSetImpl, [A._ElementCssClassSet, A.AttributeClassSet]);
-    _inheritMany(A.BoardWithPieces, [A.BoardWithPawns, A.BoardWithBishops, A.BoardWithKnights, A.BoardWithRooks, A.BoardWithKings, A.BoardWithQueens]);
     _inherit(A.ChessPiece, A.GamePiece);
     _inherit(A.EmptyPiece, A.ChessPiece);
+    _inheritMany(A.BoardWithPieces, [A.BoardWithPawns, A.BoardWithBishops, A.BoardWithKnights, A.BoardWithRooks, A.BoardWithKings, A.BoardWithQueens]);
     _mixin(A._HtmlCollection_JavaScriptObject_ListMixin, A.ListBase);
     _mixin(A._HtmlCollection_JavaScriptObject_ListMixin_ImmutableListMixin, A.ImmutableListMixin);
     _mixin(A._NodeList_JavaScriptObject_ListMixin, A.ListBase);
@@ -4934,7 +4934,7 @@
     leafTags: null,
     arrayRti: Symbol("$ti")
   };
-  A._Universe_addRules(init.typeUniverse, JSON.parse('{"PlainJavaScriptObject":"LegacyJavaScriptObject","UnknownJavaScriptObject":"LegacyJavaScriptObject","JavaScriptFunction":"LegacyJavaScriptObject","AbortPaymentEvent":"Event","ExtendableEvent":"Event","AElement":"SvgElement","GraphicsElement":"SvgElement","AudioElement":"HtmlElement","MediaElement":"HtmlElement","HtmlDocument":"Node","Document":"Node","CDataSection":"CharacterData","Text":"CharacterData","MathMLElement":"Element","HtmlFormControlsCollection":"HtmlCollection","JSBool":{"bool":[],"TrustedGetRuntimeType":[]},"JSNull":{"TrustedGetRuntimeType":[]},"JSArray":{"List":["1"],"Iterable":["1"]},"JSUnmodifiableArray":{"JSArray":["1"],"List":["1"],"Iterable":["1"]},"ArrayIterator":{"Iterator":["1"]},"JSNumber":{"num":[]},"JSInt":{"int":[],"num":[],"TrustedGetRuntimeType":[]},"JSNumNotInt":{"num":[],"TrustedGetRuntimeType":[]},"JSString":{"String":[],"TrustedGetRuntimeType":[]},"ListIterator":{"Iterator":["1"]},"MappedIterable":{"Iterable":["2"]},"MappedIterator":{"Iterator":["2"]},"WhereIterable":{"Iterable":["1"]},"WhereIterator":{"Iterator":["1"]},"Closure":{"Function":[]},"Closure2Args":{"Function":[]},"TearOffClosure":{"Function":[]},"StaticClosure":{"Function":[]},"BoundClosure":{"Function":[]},"_LinkedHashSet":{"SetBase":["1"],"LinkedHashSet":["1"],"Set":["1"],"Iterable":["1"]},"_LinkedHashSetIterator":{"Iterator":["1"]},"ListBase":{"List":["1"],"Iterable":["1"]},"SetBase":{"Set":["1"],"Iterable":["1"]},"_SetBase":{"SetBase":["1"],"Set":["1"],"Iterable":["1"]},"int":{"num":[]},"List":{"Iterable":["1"]},"Set":{"Iterable":["1"]},"Element":{"Node":[]},"HtmlElement":{"Element":[],"Node":[]},"AnchorElement":{"Element":[],"Node":[]},"AreaElement":{"Element":[],"Node":[]},"CharacterData":{"Node":[]},"_ChildrenElementList":{"ListBase":["Element"],"List":["Element"],"Iterable":["Element"],"ListBase.E":"Element"},"FormElement":{"Element":[],"Node":[]},"HtmlCollection":{"ListBase":["Node"],"ImmutableListMixin":["Node"],"List":["Node"],"JavaScriptIndexingBehavior":["Node"],"Iterable":["Node"],"ListBase.E":"Node","ImmutableListMixin.E":"Node"},"ImageElement":{"Element":[],"Node":[]},"InputElement":{"Element":[],"Node":[]},"LabelElement":{"Element":[],"Node":[]},"_ChildNodeListLazy":{"ListBase":["Node"],"List":["Node"],"Iterable":["Node"],"ListBase.E":"Node"},"NodeList":{"ListBase":["Node"],"ImmutableListMixin":["Node"],"List":["Node"],"JavaScriptIndexingBehavior":["Node"],"Iterable":["Node"],"ListBase.E":"Node","ImmutableListMixin.E":"Node"},"SelectElement":{"Element":[],"Node":[]},"_NamedNodeMap":{"ListBase":["Node"],"ImmutableListMixin":["Node"],"List":["Node"],"JavaScriptIndexingBehavior":["Node"],"Iterable":["Node"],"ListBase.E":"Node","ImmutableListMixin.E":"Node"},"_ElementCssClassSet":{"SetBase":["String"],"Set":["String"],"Iterable":["String"]},"FixedSizeListIterator":{"Iterator":["1"]},"CssClassSetImpl":{"SetBase":["String"],"Set":["String"],"Iterable":["String"]},"FilteredElementList":{"ListBase":["Element"],"List":["Element"],"Iterable":["Element"],"ListBase.E":"Element"},"AttributeClassSet":{"SetBase":["String"],"Set":["String"],"Iterable":["String"]},"SvgElement":{"Element":[],"Node":[]},"BoardWithPieces":{"ChessBoard":[]},"BoardWithPawns":{"ChessBoard":[]},"BoardWithBishops":{"ChessBoard":[]},"BoardWithKnights":{"ChessBoard":[]},"BoardWithRooks":{"ChessBoard":[]},"BoardWithKings":{"ChessBoard":[]},"BoardWithQueens":{"ChessBoard":[]},"ChessBoardView":{"ChessView":[]},"ChequeredBoard":{"ChessBoard":[]},"EmptyPiece":{"ChessPiece":[]},"DecoratorDemo":{"Demo":[]},"NoMovement":{"MovementStrategy":[]},"PawnMovement":{"MovementStrategy":[]},"KnightMovement":{"MovementStrategy":[]},"BishopMovement":{"MovementStrategy":[]},"RookMovement":{"MovementStrategy":[]},"QueenMovement":{"MovementStrategy":[]},"KingMovement":{"MovementStrategy":[]}}'));
+  A._Universe_addRules(init.typeUniverse, JSON.parse('{"PlainJavaScriptObject":"LegacyJavaScriptObject","UnknownJavaScriptObject":"LegacyJavaScriptObject","JavaScriptFunction":"LegacyJavaScriptObject","AbortPaymentEvent":"Event","ExtendableEvent":"Event","AElement":"SvgElement","GraphicsElement":"SvgElement","AudioElement":"HtmlElement","MediaElement":"HtmlElement","HtmlDocument":"Node","Document":"Node","CDataSection":"CharacterData","Text":"CharacterData","MathMLElement":"Element","HtmlFormControlsCollection":"HtmlCollection","JSBool":{"bool":[],"TrustedGetRuntimeType":[]},"JSNull":{"TrustedGetRuntimeType":[]},"JSArray":{"List":["1"],"Iterable":["1"]},"JSUnmodifiableArray":{"JSArray":["1"],"List":["1"],"Iterable":["1"]},"ArrayIterator":{"Iterator":["1"]},"JSNumber":{"num":[]},"JSInt":{"int":[],"num":[],"TrustedGetRuntimeType":[]},"JSNumNotInt":{"num":[],"TrustedGetRuntimeType":[]},"JSString":{"String":[],"TrustedGetRuntimeType":[]},"ListIterator":{"Iterator":["1"]},"MappedIterable":{"Iterable":["2"]},"MappedIterator":{"Iterator":["2"]},"WhereIterable":{"Iterable":["1"]},"WhereIterator":{"Iterator":["1"]},"Closure":{"Function":[]},"Closure2Args":{"Function":[]},"TearOffClosure":{"Function":[]},"StaticClosure":{"Function":[]},"BoundClosure":{"Function":[]},"_LinkedHashSet":{"SetBase":["1"],"LinkedHashSet":["1"],"Set":["1"],"Iterable":["1"]},"_LinkedHashSetIterator":{"Iterator":["1"]},"ListBase":{"List":["1"],"Iterable":["1"]},"SetBase":{"Set":["1"],"Iterable":["1"]},"_SetBase":{"SetBase":["1"],"Set":["1"],"Iterable":["1"]},"int":{"num":[]},"List":{"Iterable":["1"]},"Set":{"Iterable":["1"]},"Element":{"Node":[]},"HtmlElement":{"Element":[],"Node":[]},"AnchorElement":{"Element":[],"Node":[]},"AreaElement":{"Element":[],"Node":[]},"CharacterData":{"Node":[]},"_ChildrenElementList":{"ListBase":["Element"],"List":["Element"],"Iterable":["Element"],"ListBase.E":"Element"},"FormElement":{"Element":[],"Node":[]},"HtmlCollection":{"ListBase":["Node"],"ImmutableListMixin":["Node"],"List":["Node"],"JavaScriptIndexingBehavior":["Node"],"Iterable":["Node"],"ListBase.E":"Node","ImmutableListMixin.E":"Node"},"ImageElement":{"Element":[],"Node":[]},"InputElement":{"Element":[],"Node":[]},"LabelElement":{"Element":[],"Node":[]},"_ChildNodeListLazy":{"ListBase":["Node"],"List":["Node"],"Iterable":["Node"],"ListBase.E":"Node"},"NodeList":{"ListBase":["Node"],"ImmutableListMixin":["Node"],"List":["Node"],"JavaScriptIndexingBehavior":["Node"],"Iterable":["Node"],"ListBase.E":"Node","ImmutableListMixin.E":"Node"},"SelectElement":{"Element":[],"Node":[]},"_NamedNodeMap":{"ListBase":["Node"],"ImmutableListMixin":["Node"],"List":["Node"],"JavaScriptIndexingBehavior":["Node"],"Iterable":["Node"],"ListBase.E":"Node","ImmutableListMixin.E":"Node"},"_ElementCssClassSet":{"SetBase":["String"],"Set":["String"],"Iterable":["String"]},"FixedSizeListIterator":{"Iterator":["1"]},"CssClassSetImpl":{"SetBase":["String"],"Set":["String"],"Iterable":["String"]},"FilteredElementList":{"ListBase":["Element"],"List":["Element"],"Iterable":["Element"],"ListBase.E":"Element"},"AttributeClassSet":{"SetBase":["String"],"Set":["String"],"Iterable":["String"]},"SvgElement":{"Element":[],"Node":[]},"EmptyPiece":{"ChessPiece":[]},"ChessBoardView":{"ChessView":[]},"BoardWithPieces":{"ChessBoard":[]},"BoardWithPawns":{"ChessBoard":[]},"BoardWithBishops":{"ChessBoard":[]},"BoardWithKnights":{"ChessBoard":[]},"BoardWithRooks":{"ChessBoard":[]},"BoardWithKings":{"ChessBoard":[]},"BoardWithQueens":{"ChessBoard":[]},"ChequeredBoard":{"ChessBoard":[]},"NoMovement":{"MovementStrategy":[]},"PawnMovement":{"MovementStrategy":[]},"KnightMovement":{"MovementStrategy":[]},"BishopMovement":{"MovementStrategy":[]},"RookMovement":{"MovementStrategy":[]},"QueenMovement":{"MovementStrategy":[]},"KingMovement":{"MovementStrategy":[]},"DecoratorDemo":{"Demo":[]}}'));
   A._Universe_addErasedTypes(init.typeUniverse, JSON.parse('{"_SetBase":1}'));
   var type$ = (function rtii() {
     var findType = A.findType;
